requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
amqp==5.2.0; python_version >= '3.6'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
av==10.0.0
billiard==4.2.0; python_version >= '3.7'
boto3==1.34.30; python_version >= '3.8'
botocore==1.34.130; python_version >= '3.8'
cachetools==5.3.3; python_version >= '3.7'
celery[redis,sqs]==5.4.0; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cobble==0.1.4; python_version >= '3.5'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cssselect==1.2.0; python_version >= '3.7'
ctranslate2==3.24.0; python_version >= '3.8'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
django==4.2.3; python_version >= '3.8'
django-celery-results==2.5.1
django-cleanup==8.1.0
django-cors-headers==4.2.0; python_version >= '3.8'
django-easy-health-check==1.0.1; python_version >= '3.6'
django-environ==0.10.0; python_version >= '3.5' and python_version < '4'
django-filter==23.3; python_version >= '3.7'
django-ordered-model==3.7.4
django-storages==1.14.2; python_version >= '3.7'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.2.2; python_version >= '3.7'
drf-spectacular==0.26.5; python_version >= '3.6'
exceptiongroup==1.2.1; python_version < '3.11'
faster-whisper==0.7.1; python_version >= '3.8'
filelock==3.15.3; python_version >= '3.8'
flatbuffers==24.3.25
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.0; python_version >= '3.8'
gevent==23.9.1; python_version >= '3.8'
google-api-core[grpc]==2.19.0; python_version >= '3.7'
google-auth==2.30.0; python_version >= '3.7'
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-translate==3.12.1; python_version >= '3.7'
googleapis-common-protos==1.63.1; python_version >= '3.7'
greenlet==3.0.3; python_version < '3.11' and platform_python_implementation == 'CPython'
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0; python_version >= '3.7'
html2text==2020.1.16; python_version >= '3.5'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
huggingface-hub==0.23.4; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.7; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu[sqs]==5.3.7; python_version >= '3.8'
langchain==0.1.14; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.38; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.52; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.1; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-text-splitters==0.0.2; python_version < '4.0' and python_full_version >= '3.8.1'
langdetect==1.0.9
langsmith==0.1.81; python_version < '4.0' and python_full_version >= '3.8.1'
lxml[html_clean]==5.2.2; python_version >= '3.6'
lxml-html-clean==0.1.1
mammoth==1.7.1; python_version >= '3.7'
markdown==3.5.2; python_version >= '3.8'
marshmallow==3.21.3; python_version >= '3.8'
mpmath==1.3.0
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
mysqlclient==2.1.1; python_version >= '3.5'
nltk==3.8.1; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
onnxruntime==1.18.0
openai==1.13.3; python_full_version >= '3.7.1'
orjson==3.10.5; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pgvector==0.2.5; python_version >= '3.8'
playwright==1.41.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==4.25.3; python_version >= '3.8'
psycopg[binary,pool]==3.1.16; python_version >= '3.7'
psycopg-binary==3.1.16
psycopg-pool==3.2.2
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycurl==7.45.3
pydantic==2.7.4; python_version >= '3.8'
pydantic-core==2.18.4; python_version >= '3.8'
pydub==0.25.1
pyee==11.0.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pypdfium2==4.24.0; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.1; python_version >= '3.6'
readability-lxml==0.8.1
redis==5.0.6
referencing==0.35.1; python_version >= '3.8'
regex==2023.8.8; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
rpds-py==0.18.1; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
s3transfer==0.10.1; python_version >= '3.8'
scikit-learn==1.4.1.post1; python_version >= '3.9'
scipy==1.13.1; python_version >= '3.9'
sentry-sdk[celery,django]==1.41.0
setuptools==70.1.0; python_version >= '3.8'
shortuuid==1.0.11; python_version >= '3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.31; python_version >= '3.7'
sqlparse==0.4.4; python_version >= '3.5'
sympy==1.12.1; python_version >= '3.8'
tenacity==8.4.1; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
tiktoken==0.6.0; python_version >= '3.8'
tokenizers==0.13.3
tqdm==4.66.4; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
unidecode==1.3.7; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.6'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
yarl==1.9.4; python_version >= '3.7'
youtube-transcript-api==0.6.2
zope.event==5.0; python_version >= '3.7'
zope.interface==6.4.post2; python_version >= '3.7'
